+++ src_test/test_ppx_deriving.ml
@@ -1,9 +1,13 @@
 open OUnit2
 
-let test_shorthand ctxt =
+let test_inline ctxt =
   let sort = List.sort [%derive.ord: int * int] in
   assert_equal ~printer:[%derive.show: (int * int) list]
                [(1,1);(2,0);(3,5)] (sort [(2,0);(3,5);(1,1)])
+
+let test_inline_shorthand ctxt =
+  assert_equal ~printer:(fun x -> x)
+               "[(1, 1); (2, 0)]" ([%show: (int * int) list] [(1,1); (2,0)])
 
 let suite = "Test ppx_deriving" >::: [
     Test_deriving_show.suite;
@@ -13,7 +17,8 @@ let suite = "Test ppx_deriving" >::: [
     Test_deriving_iter.suite;
     Test_deriving_map.suite;
     Test_deriving_fold.suite;
-    "test_shorthand" >:: test_shorthand;
+    "test_inline"           >:: test_inline;
+    "test_inline_shorthand" >:: test_inline_shorthand;
   ]
 
 let _ =